Insert space between period and non-whitespace
In the tesserae corpus, periods were found to be succeeded
by all of the following characters (in regex).
\s—\-†,:\"\'“”‘’\]\.;&vâ\*\^\}（）\dα-ωΑ-Ὠ`̔
Note that the omega is followed by COMBINING COMMA ABOVE,
followed by GREEK VARIA, followed by COMBINING REVERSED
COMMA ABOVE. The COMBINING characters render above the
preceding character, not after.

Sometimes the character that succeeds a period is not meant to
be there due to a parsing bug. The following are examples (regex):
“‘—\-†&vâ\*\^（α-ωΑ-Ὠ`̔
If there is no space between a period and a succeeding word, opening
quote, ampersand, etc. this is almost certainly a mistake.

Sometimes the character that succeeds a period is not a mistake.
The following are examples (regex)
\s\d”’\"\'\}\]）\.,:;
Periods are supposed to be followed by whitespaces, closing quotes,
closing brackets, etc.

I inserted a space between each period followed by a mistaken character.
Here are the steps I took.
In Sublime Text, I pressed command-shift-f (⌘ ⇧ f).
I enabled regular expressions with option-command-r (⌥ ⌘ r).
Find: (?!\..*?>)\.(?=[“‘—\-†&vâ\*\^（α-ωΑ-Ὠ`̔])
Where: -*.venv/*,*tesserae/texts/grc/*,-*.pickle
Replace: .
(The replace field is a period followed by a space)
Then I pressed the "Replace" button.