Fix out-of-order lines
The following lines were out of order, appearing after l. 290:
```
<sil. 11.453> namque chaos, caecam quondam sine sidere molem
<sil. 11.454> non surgente die, ac mundum sine luce canebat.
<sil. 11.455> tum deus ut liquidi disclusset stagna profundi
<sil. 11.456> tellurisque globum media compage locasset;
<sil. 11.457> ut celsum superis habitare dedisset Olympum,
<sil. 11.458> castaque Saturni monstrabat saecula patris.
```

This change was a suggestion of Summers, but is rejected by Delz (the modern edition). I have no strong opinion on the textual issue, but I think that if the lines are accepted as a reordering then they should be renumbered, because it can cause errors in automatic processing of results.